ATTENTION!!! (realted to previous commit)
I accidently stages new changes(of new commit i was going to make), then ammended the commit message(of previous commit), so the new changes also got commited with the previous commit changes.

previous commit was about just connecting app.js with web3 and metamask. but now it has code of making seprate component for Navbar and adding identicon next to the address of user (loaded address, and then used state component to display it).